Add -P switch used to specify mbed-ls pooling timeout <sec>
This feature also allows us to specify timeout for:
* mount point
* serial port mount

Flow:

1. Start mbedhtrun
2. Pooling for mount point of device (-t <target_id>) using mbed-ls.
	* Timeout == 60 sec (or changed with -P)
3. Pooling for serial port of device (-t <target_id>) using mbed-ls.
	* Timeout == 60 sec (or changed with -P)

This means -P <sec> defines the same timeout for both poling operations.
This can be changed, fragemented in the future.

In code default value is set to 60 secons but switch -P can exactly be used to
change this timeout depending on application.