Replace lockfile context mgr usage with lock.aquire() and lock.release()
To catch LockTImeout exception we can't use context mgr. We need to use explicit
calls to acquire(timeout=1) and release() to catch timeout.
TImeout may occur on some Resource Managers instance.

LockFile objects support the context manager protocol used by the statement:with
statement. The timeout option is not supported when used in this fashion. While
support for timeouts could be implemented, there is no support for handling the
eventual Timeout exceptions raised by the __enter__() method, so you would have
to protect the with statement with a try statement. The resulting construct
would not be any simpler than just using a try statement in the first place.