Fixing issue with multiprocess Queue
This commit fixes an intermittent failure with the multiprocess Queue used
within the host test. There is a chance that a Queue will return that it
is empty even if an item has been added to it (see the Queue documentation
here:
https://docs.python.org/2/library/multiprocessing.html#pipes-and-queues).
This issue intermittently arrises in the following scenario:

- A DUT sends the {{end;success}} KV pair
- A DUT sends the {{__exit;0}} KV pair
- The {{end;success}} KV pair is pulled from the queue, which puts a
  callback for notify_complete to be called (after the __exit KV pair)
- The {{__exit;0}} KV pair is pulled from the queue, which breaks out of
  the loop
- The queue is checked to see if it is empty. This is the intermittent
  point of failure, as this sometimes returns True even though there
  should be a notify_compelte KV pair in the queue
- Since the queue is presumed to be empty, the result isn't properly
  updated and the test fails incorrectly

The fix for this is not to break out of the queue loop on __exit or any
other KV pair or error condition. Instead, replace these 'break'
statements with another queue event called '__exit_event_queue'. This is
the only event that will call 'break'. The rule for this event is that it
will always be the last event in the queue. Since the Queue.empty()
function cannot be trusted, this is used instead to mark the end of the
host test execution.

This solution was inspired by the blog post here:
http://bryceboe.com/2011/01/28/the-python-multiprocessing-queue-and-large-objects/