Adding 'write_timeout' to serial connection
Writing to a Serial port works differently on Linux than on Windows. The
pySerial documentation says that the serial.write() function will default
to blocking if 'write_timeout` is not set. On Linux, this means that the
serial.write blocks forever if the device has not received the data (which
happens in certain cases for different boards). This also ocurrs in
Windows, however it appears that Windows has a relatively large internal
buffer (~10K) that will cause the serial.write() function to return
immediately if there is still space in the buffer. Only when you send
large amounts of data does the function block.

This commit attempts to standardize the behavior of the host test (and to
prevent indefinitely hangs) by introducing a default write timeout of 5
seconds.