Searching project recursively for target information
This commit makes greentea perform a directory walk in search of a
targets.json file. Previously, there were only a few hardcoded locations
that were searched for this file. This operation is more expensive
unfortunately, however it is a far more robust solution. It no longer
makes assumptions on the project's structure. It should be resilient against
any more changes made to the structure of mbed OS.