Fix issue were a method is attempted to be called on a NoneType.
In get_mbeds() if the regular expression for looking for a tid fails then
a NoneType occurs. The code then attempts to call the group() function on
this which causes an exception.
This fix adds a check for the regular expression returning a NoneType
before trying to call the method.