Fixes issue #198 where an error occurs when trying to use the endswith() method on a NoneType.
On windows 7 if a target was detected but for some reason the serial port
was not, then mbed-ls would return a NoneType for the serial port field in
the ready_mbed_devices list. This serial port field under normal
circumstances will contain a string and has the endswith() method called
on it. This is however illegal for a NoneType.
The fix is to read the serial port field first and test whether or not it
is the NoneType prior to calling the method. If it is the NoneType then
report an error and skip this target.