Removes the need to strip lines and fixes timeout issue.
This change stops stripping the received lines, making the saved
output look closer to the actual output you see in a serial terminal.

It also fixes an issue when comparing an example output where the
timeouts are set differently. This is because when the loop enters the
"cosume_premable_events" part, it adds an extra 60 seconds to execution
time. This was only being done when "serial_output_file" was set, now
it's being done when "compare_log" is set as well (by changing the
if/else structure).