Fix error in regbin2str() in python3
The python3 version of mbed-ls was breaking under windows because the
regbin2str() function was returning a different type for each version of python.
The filter function returns a list in python2, which seems to be implicitly
converted to a string, in python3 filter returns a filter object.

This is fixed by explicitly converting the filter to a string before returning.