Windows API can now tell if mbeds are actually plugged in
The Windows registry is responsible for keeping track of all USB
devices. Previously, we were using a portion of the registry that is
always cached, even if a device is removed. This caused certain network
drives to appear as mbeds if they attached to a previously attached
mbed's drive letter (network drives don't appear to update the registry).

Now we are checking the enumeration part of the registry to ensure the
device is actually plugged in before return it as a result.