Avoid deadlock on reading PList process output
On Mac (Darwin), we read much of the USB and MSD info from XML Plists.
These plists are parsed by python's `plistlib`. The prior implementation
would assume that the process generating the plist would complete by the
time the XML parser reached the end of it's current stream. As it turns
out, python3's XML parser violates this assumption. This change forces
the plist-generating process to complete before parsing starts. This may
be a tad slower, but it makes sure that we don't enter a state where we
have finished processing a plist, and the plist-generating process has
filled it's stdout pipe and is also blocked on mbed-ls reading its output.