ci: Add deploy job to release to PyPI
The release to PyPI is now automated. The trigger is pushing a tag
(vX.X.X) to master.

A new deploy job has been added to azure pipelines which will be
triggered when a tag is pushed to master. The deploy job will check
latest tag with the previous tag and will make a release to PyPI if
there are changes.