htrun: serial: Use provided serial port when given
In order to work well with boards via the pyocd copy method, we should
not attempt to list Mbed style boards to see if the serial port changed
names. We should instead use the serial port as provided on the command
line. If we attempt to list Mbed style boards, we will run into an error
when the board is not Mbed-style, even if the command-line provided
serial port is correct. To add insult to injury, the error message given
in such cases gives advice that is impossible to follow; the output from
htrun recommends you provide the -u option, but you cannot; the -u
option is for mbed-ls, not htrun.

    $ mbedhtrun -f build/test.elf -p /dev/ttyUSB1:115200 -r pyocd -c pyocd --target-id 0240
    [1627903348.60][mbedls.lstools_base]MBED with target id '0240000034544e45001e00048e3800515a91000097969900' is connected, but not mounted. Use the '-u' flag to include it in the list.

The reason we are not simply overriding the pyocd plugin's
check_serial_port_ready() method is because we don't want to attempt
auto-detection of the serial port for any type of device, not just those
we'd use pyocd with. We always want to use the provided serial port when
passed in. If users want auto-detection of the serial port to use, they
can avoid providing the port on the command line to get the previous
behavior.

Fixes #267